pyproject.toml
[project]
name = "vic3-analysis"
authors = [{ name = "Brian Li", email = "brianlizsy@qq.com" }]
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pyradox-txt-parser>=5.0.1",
    "scipy>=1.17.1",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]


[dependency-groups]
dev = [
    "markdown-callouts>=0.4.0",
    "mkdocs-material>=9.7.4",
    "mkdocstrings[python]>=1.0.3",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
]
